Port remaining information_schema rust tests to sqllogictests
Closes: #6301
