+++ datafusion/core/tests/sql/mod.rs
@@ -83,7 +83,6 @@ pub mod create_drop;
 pub mod explain_analyze;
 pub mod expr;
 pub mod group_by;
-pub mod information_schema;
 pub mod joins;
 pub mod limit;
 pub mod order;
